Fixed make errors

+++ src/send_response.c
@@ -35,7 +35,7 @@ void send_response(int incoming_fd, char *path)
   }
 
   int file_path_size = snprintf(NULL, 0, "%s%s", ASSET_DIR, path);
-  char *file_path = malloc(file_path_size + 1); // yay null terminator
+  char *file_path = malloc((size_t)(file_path_size + 1)); // yay null terminator
   snprintf(file_path, (size_t)(file_path_size + 1), "%s%s", ASSET_DIR, path);
 
   // libmagic didn't work, so here's a temp? solution...
@@ -47,7 +47,7 @@ void send_response(int incoming_fd, char *path)
   if (extension == NULL)
   {
     file_path_size += 5;
-    file_path = realloc(file_path, file_path_size + 1); // yay null terminator
+    file_path = realloc(file_path, (size_t)(file_path_size + 1)); // yay null terminator
     snprintf(file_path, (size_t)(file_path_size + 1), "%s%s.html", ASSET_DIR, path);
     extension = ".html";
   }
@@ -80,7 +80,7 @@ void send_response(int incoming_fd, char *path)
     // TODO: ALso please find better names and don't hardcode these values
     printf("File not found\n");
     char *response = "HTTP/1.0 404 Not Found\r\nContent-Length: 826\r\nContent-Type: text/html\r\n";
-    send_data(incoming_fd, response, strlen(response));
+    send_data(incoming_fd, response, (long)strlen(response));
     int data_404_size = 826;
     FILE *file_404 = fopen("./pages/404.html", "rb");
     unsigned char file_data_404[data_404_size];
